+++ fhtable/fhtable.h
@@ -47,6 +47,6 @@ unsigned long fhtable_move(struct fhtable *dest, struct fhtable *src,
 /*
  * Dumps the hash table to stdout for debugging purpose.
  */
-void fhtable_dump(const struct fhtable *ht, int level);
+void fhtable_dump(const struct fhtable *ht, unsigned long level);
 
 #endif
